fix: sửa authorization cho websocket và các phương thức trong groupChatServiceImpl

+++ src/main/java/com/e_messenger/code/service/impl/GroupChatServiceImpl.java
@@ -80,9 +80,12 @@ public class GroupChatServiceImpl extends GroupChatService {
     }
 
     @Override
-    @PreAuthorize("@participantUtil.hasManagementRole(@conversationQueryServiceImpl.getConversationById(#groupId), @userService.getCurrentUser())")
     public Conversation addParticipants(String groupId, List<String> participantIds) {
-        Conversation group = queryService.getConversationById(groupId);
+        User curUser = userService.getCurrentUser();
+        Conversation group = queryService.getConversationById(groupId, curUser.getId());
+
+        if(!participantUtil.hasManagementRole(group, curUser))
+            throw new AppException(StatusCode.UNCATEGORIZED);
 
         LinkedHashSet<User> curUserList = new LinkedHashSet<>(participantUtil.toUsers(group.getParticipants()));
         List<User> validUsers = userUtil.getValidUsers(participantIds);
@@ -99,9 +102,12 @@ public class GroupChatServiceImpl extends GroupChatService {
     }
 
     @Override
-    @PreAuthorize("@participantUtil.canAffect(@conversationQueryServiceImpl.getConversationById(#groupId), @userService.getCurrentUser(), #removeIds)")
     public Conversation removeParticipants(String groupId, List<String> removeIds) {
-        Conversation group = queryService.getConversationById(groupId);
+        User curUser = userService.getCurrentUser();
+        Conversation group = queryService.getConversationById(groupId, curUser.getId());
+
+        if(!participantUtil.canAffect(group, curUser, removeIds))
+            throw new AppException(StatusCode.UNCATEGORIZED);
 
         group.setParticipants(
                 removeAll(group.getParticipants(), removeIds)
@@ -113,33 +119,36 @@ public class GroupChatServiceImpl extends GroupChatService {
     }
 
     @Override
-    @PreAuthorize("@participantUtil.hasRole(" +
-            "@conversationQueryServiceImpl.getConversationById(#groupId), " +
-            "@userService.getCurrentUser(), " +
-            "T(com.e_messenger.code.entity.enums.ConversationRole).OWNER)")
     public void deleteGroup(String groupId) {
-        Conversation group = conversationQueryService.getConversationById(groupId);
+        User curUser = userService.getCurrentUser();
+        Conversation group = conversationQueryService.getConversationById(groupId, curUser.getId());
+
+        if(!participantUtil.hasRole(group, curUser, ConversationRole.OWNER))
+            throw new AppException(StatusCode.UNCATEGORIZED);
+
         conversationRepo.delete(group);
     }
 
     @Override
-    @PreAuthorize("@participantUtil.hasRole(" +
-            "@conversationQueryServiceImpl.getConversationById(#groupId), " +
-            "@userService.getCurrentUser(), " +
-            "T(com.e_messenger.code.entity.enums.ConversationRole).OWNER)")
     public Conversation setOwner(String groupId, String newOwnerId) {
-        Conversation group = conversationQueryService.getConversationById(groupId);
+        User curUser = userService.getCurrentUser();
+        Conversation group = conversationQueryService.getConversationById(groupId, curUser.getId());
+
+        if(!participantUtil.hasRole(group, curUser, ConversationRole.OWNER))
+            throw new AppException(StatusCode.UNCATEGORIZED);
+
         participantUtil.changeOwner(group, userService.getCurrentUser().getId(), newOwnerId);
         return conversationRepo.save(group);
     }
 
     @Override
-    @PreAuthorize("@participantUtil.hasRole(" +
-            "@conversationQueryServiceImpl.getConversationById(#groupId), " +
-            "@userService.getCurrentUser(), " +
-            "T(com.e_messenger.code.entity.enums.ConversationRole).OWNER)")
     public Conversation setCoOwner(String groupId, List<String> coOwnerIds) {
-        Conversation group = conversationQueryService.getConversationById(groupId);
+        User curUser = userService.getCurrentUser();
+        Conversation group = conversationQueryService.getConversationById(groupId, curUser.getId());
+
+        if(!participantUtil.hasRole(group, curUser, ConversationRole.OWNER))
+            throw new AppException(StatusCode.UNCATEGORIZED);
+
         List<User> validUsers = userUtil.getValidUsers(coOwnerIds);
 
         participantUtil.setCoOwners(group, validUsers);
@@ -147,9 +156,12 @@ public class GroupChatServiceImpl extends GroupChatService {
     }
 
     @Override
-    @PreAuthorize("@participantUtil.canAffect(@conversationQueryServiceImpl.getConversationById(#groupId), @userService.getCurrentUser(), #participantIds)")
     public Conversation toMember(String groupId, List<String> participantIds) {
-        Conversation group = conversationQueryService.getConversationById(groupId);
+        User curUser = userService.getCurrentUser();
+        Conversation group = conversationQueryService.getConversationById(groupId, curUser.getId());
+
+        if(!participantUtil.canAffect(group, curUser, participantIds))
+            throw new AppException(StatusCode.UNCATEGORIZED);
 
         List<User> validUsers = userUtil.getValidUsers(participantIds);
         participantUtil.setMembers(group, validUsers);
@@ -159,7 +171,7 @@ public class GroupChatServiceImpl extends GroupChatService {
     @Override
     public boolean leaveConversation(String groupId) {
         User curUser = userService.getCurrentUser();
-        Conversation group = queryService.getConversationById(groupId);
+        Conversation group = queryService.getConversationById(groupId, curUser.getId());
 
         if(group.getParticipants().size() == 1)
             throw new AppException(StatusCode.UNCATEGORIZED);
